Update of the project structure and addition of workflows to build the project & build an unsigned debug or release apk

+++ app/src/main/java/cu/axel/smartdock/utils/DeviceUtils.java
@@ -130,7 +130,7 @@ public class DeviceUtils {
         try {
             Method getUserIcon=um.getClass().getMethod("getUserIcon", int.class);
             Method myUserId=UserHandle.class.getMethod("myUserId");
-            int id=myUserId.invoke(UserHandle.class);
+            int id=(int)myUserId.invoke(UserHandle.class);
             userIcon = (Bitmap) getUserIcon.invoke(um, id);
             if (userIcon != null)
                 userIcon = Utils.getCircularBitmap(userIcon);
